requirements.txt
# Core API Framework
fastapi==0.120.0
uvicorn[standard]==0.34.0
python-multipart==0.0.9

# OpenAI Integration
openai==2.6.0

# File Processing
aiofiles>=23.2.1
PyPDF2>=3.0.0
python-docx>=1.1.0
lxml==6.0.2
Jinja2>=3.1.0  # Templating for prompt engineering

# Configuration and Environment
python-dotenv==1.1.1
pydantic==2.12.3
pydantic-settings==2.0.3
pydantic_core==2.41.4

# HTTP and Networking
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
certifi==2025.10.5
idna==3.11
anyio==4.11.0
sniffio==1.3.1

# Utilities
click==8.3.0
tqdm==4.67.1
distro==1.9.0
jiter==0.11.1
typing_extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0

# Build Tools
build==1.3.0
pip-tools==7.5.1
setuptools==80.9.0
wheel==0.45.1
packaging==25.0
pyproject_hooks==1.2.0

# Testing Dependencies (Optional - for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
